refactor(viewmodel): move VM implementation complexity in repository.
