+++ openfeedback/src/commonMain/kotlin/io/openfeedback/extensions/Project.ext.kt
@@ -1,0 +1,5 @@
+package io.openfeedback.extensions
+
+import io.openfeedback.model.Project
+
+internal fun Project.commentVoteItemId(): String? = voteItems.find { it.type == "text" }?.id
