+++ openfeedback/src/commonMain/kotlin/io/openfeedback/model/FirestoreModels.kt
@@ -1,6 +1,5 @@
 package io.openfeedback.model
 
-import kotlinx.datetime.Clock
 import kotlinx.datetime.Instant
 import kotlinx.serialization.Serializable
